+++ HISTORY.rst
@@ -1,8 +1,8 @@
 Release History
 ===============
 
-0.6.0 (?)
----------
+0.6.0 (2014-08-09)
+------------------
 
 * Implemented ``serialports monitor`` (`issue #10 <https://github.com/ivankravets/platformio/issues/10>`_)
 * Fixed an issue ``ImportError: No module named platformio.util`` (`issue #9 <https://github.com/ivankravets/platformio/issues/9>`_)